+++ backend/src/controllers/reports/genericReportController.js
@@ -1,0 +1,143 @@
+// Libraries
+const useCommonComponents = require("../useCommonComponents/index");
+const useController = require("../useController/index");
+const cdogs = useCommonComponents("cdogs");
+
+// Utilities
+const {
+  getDocumentApiBody,
+  getReportAndSetRequestHeaders,
+  pdfConfig,
+  validateQueryParameters,
+  getCurrentDate,
+} = require("./helpers");
+
+// Constants
+const what = { single: "report", plural: "reports" };
+
+/**
+ *
+ * USING THIS CONTROLLER:
+ *
+ * How to change the route: (pseudocode)
+ * EXAMPLE ROUTE: backend/src/routes/reports/Tab_50_rpt_PF_NetRecoverySummaryByQuarter.js
+ *
+ * const getControllerFrom = require("@controllers/reports/genericReportController");
+ * const name = `{whatever your report name is}`;
+ * const controller = getControllerFrom(name);
+ *
+ * How to change the model: (export the data instead of the query - pseudocode)
+ * EXAMPLE MODEL: backend/src/models/reports/Tab_50_rpt_PF_NetRecoverySummaryByQuarter.js
+ * let PSEUDOCODE_EXAMPLE = async (query) => {
+ * try {
+ * const { <args> } = query; // this could be fiscal, portfolio, date, etc.
+ *
+ * const [[{ fiscal_year }], report, report_totals] = await Promise.all([
+ * // these queries are defined in the model, and return the knex.js queries.
+ * queries.report(<some arg>),
+ * queries.totals(<some arg>),
+ * ]);
+ *
+ * return {
+ * report,
+ * report_totals,
+ * };
+ * } catch (error) {
+ * console.error(`Model error!: query parameter received: ${JSON.stringify(query)}`);
+ * console.error(`*** ${error} **** : returning NULL!.`);
+ * return null;
+ * }
+ * };
+ */
+
+/**
+ * Retrieves a controller from the specified filename.
+ *
+ * @param   {string} filename - The name of the file.
+ * @returns {object}          - The controller object.
+ */
+const getControllerFrom = (filename) => {
+  const model = require(`@models/reports/${filename}`);
+  const controller = useController(model, what);
+
+  const getReport = getReportAndSetRequestHeaders();
+
+  /**
+   * Handles the generation and delivery of a report.
+   *
+   * @async
+   * @param   {object}                 request - The request object containing information for generating the report.
+   * @param   {object}                 reply   - The reply object used to send the response.
+   * @returns {Promise<object | null>}         - A Promise that resolves to the generated report data or null if an error occurred.
+   * @throws {Error} - If an unexpected error occurs during the report generation process.
+   */
+  const reportHandler = async (request, reply) => {
+    controller.userRequires(request, "PMO-Reports-Capability", reply);
+
+    try {
+      const { query } = request;
+      const { templateType } = validateQueryParameters(query);
+      const modifiedGetReport = getReportAndSetRequestHeaders(templateType);
+      controller.getReport = modifiedGetReport;
+
+      const result = (await getDataFromModel(query, model)) ?? null;
+
+      if (null === result) {
+        reply.code(404);
+        return controller.noQuery(reply, `There was a problem looking up this Report.`);
+      } else {
+        await sendToCdogs({ result, filename, templateType, request });
+        return result;
+      }
+    } catch (err) {
+      reply.code(500);
+      return controller.failedQuery(reply, err, what);
+    }
+  };
+
+  return {
+    ...controller,
+    getReport,
+    [filename]: reportHandler,
+  };
+};
+
+/**
+ * Retrieves data from a model based on fiscal or portfolio information.
+ *
+ * @param   {string|Array[string]|number} query - The parameters for retrieving data.
+ * @param   {object}                      model - The model to retrieve data from.
+ * @returns {Promise}                           - A promise that resolves with the retrieved data.
+ */
+const getDataFromModel = async (query, model) => {
+  /* The query we are passing to the model is a list of
+   * query parameters we forward to the route from the front end dropdown menus.
+   * most reports will either have query.fiscal (the fiscal year), query.date (date for the report period)
+   *  or query.portfolio (portfolio for financial reports that summarize expenses costs, and recoveries)
+   */
+  const result = await model.getAll(query);
+  return result
+    ? {
+        date: await getCurrentDate(),
+        ...result,
+      }
+    : null;
+};
+
+/**
+ * Sends the result data to Cdogs for rendering a template.
+ *
+ * @param   {object}  params              - The parameters for sending to Cdogs.
+ * @param   {any}     params.result       - The result data to be sent.
+ * @param   {string}  params.filename     - The name of the file.
+ * @param   {string}  params.templateType - The type of the template.
+ * @param   {object}  params.request      - The request object.
+ * @returns {Promise}                     - A promise that resolves when the data is sent to Cdogs.
+ */
+const sendToCdogs = async ({ result, filename, templateType, request }) => {
+  const templateFileName = `${filename}.${templateType}`;
+  const body = await getDocumentApiBody(result, templateFileName, templateType);
+  request.data = await cdogs.api.post("/template/render", body, pdfConfig);
+};
+
+module.exports = getControllerFrom;